add vw_qgep_channel
At the moment, qgep creates one channel per reach. assigning one channel to multiple reaches is rather bothersome (select both reaches and alter to join the channels of a reachIn order to select channels by clicking on a reach it is advisable to add a view vw_qgep_channel onto which we can then build a reference.